tests/wpt/tests/svg/text/parsing/text-decoration-style-computed.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml">
  <title>SVG Text CSS Decoration Test: getComputedStyle().textDecorationStyle</title>
  <metadata>
    <h:link rel="help" ref="https://drafts.csswg.org/css-text-decor-3/#text-decoration-style-property"/>
    <h:meta name="assert" content="SVG text-decoration-style computed value is as specified."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/computed-testcommon.js"/>
  <script><![CDATA[

test_computed_value("text-decoration-style", "solid");
test_computed_value("text-decoration-style", "double");
test_computed_value("text-decoration-style", "dotted");
test_computed_value("text-decoration-style", "dashed");
test_computed_value("text-decoration-style", "wavy");

  ]]></script>
</svg>
